Last Changes
Issues:
Other player does not see other player's revealed cells (can't see each others' updated boards): I think this is because I cannot do socket.send(JSON.stringify({data}))
properly, or im not sure how to do it (i'll look tmr)

The gameboard does not reveal the cell when you click on it, but reveals it when you put a flag somewhere with right click: I think this is because of the way I declared the boardData variable, I declared it as a constant. I know it would be better to create a let variable and initialize the board it in a function. So I believe that the board "resets" everytime so it never shows the board. But this might be wrong too because when you click a cell you clicked before, it knows you already revealed that.

You don't get you win, you lost alerts: I think this is also related the socket.send function and I am not properly using it.

The reset function doesnt rly work for some reason.